more PR changes
