+++ rundeckapp/grails-spa/packages/ui-trellis/src/app/components/common/types/commonTypes.ts
@@ -15,10 +15,4 @@ type ModalButtons = {
   messageCode?: string;
 };
 
-interface JobDetailsData {
-  jobName: string;
-  description?: string;
-  groupPath?: string;
-}
-
-export { ModalLinks, ModalButtons, JobDetailsData };
+export { ModalLinks, ModalButtons };
