+++ rundeckapp/grails-spa/packages/ui-trellis/src/app/components/job/details/tests/DetailsEditor.spec.ts
@@ -1,6 +1,6 @@
 import { shallowMount, VueWrapper } from "@vue/test-utils";
 import DetailsEditor from "../DetailsEditor.vue";
-import DetailsGroupModal from "../DetailsGroupModal.vue";
+import DetailsGroupModal from "../../../common/DetailsGroupModal.vue";
 import ScheduledExecutionDetails from "../../../common/ScheduledExecutionDetails.vue";
 import AceEditor from "@/library/components/utils/AceEditor.vue";
 import UiSocket from "@/library/components/utils/UiSocket.vue";
