+++ rundeckapp/grails-spa/packages/ui-trellis/src/app/components/job/details/types/detailsType.ts
@@ -1,0 +1,7 @@
+interface JobDetailsData {
+  jobName: string;
+  description?: string;
+  groupPath?: string;
+}
+
+export { JobDetailsData };
